+++ mywords-go/server/server.go
@@ -245,10 +245,10 @@ func (s *Server) restoreFromBackUpDataKnownWordsFile(f *zip.File) error {
 		return err
 	}
 	// merge s.knownWordsMap and knownWordsMap
-	for k, v := range knownWordsMap {
-		for k1, v1 := range v {
+	for firstLetter, v := range knownWordsMap {
+		for word, level := range v {
 			//s.knownWordsMap[k][k1] = v1
-			s.knownWordsMap.Set(k, k1, v1)
+			s.knownWordsMap.Set(firstLetter, word, level)
 		}
 	}
 	return s.saveKnownWordsMapToFile()
@@ -411,13 +411,13 @@ func (s *Server) parseChartDateLevelCountMapFromGobFile(r io.ReadCloser) error {
 			//	s.chartDateLevelCountMap[date][level] = make(map[string]struct{})
 			//}
 			for word := range wordMap {
-				mData, ok := s.chartDateLevelCountMap.Get(date, level)
-				if !ok {
-					mData = make(map[string]struct{})
+				wp, _ := s.chartDateLevelCountMap.Get(date, level)
+				wordMapNew := make(map[string]struct{}, len(wp))
+				for k, v := range wp {
+					wordMapNew[k] = v
 				}
-				mData[word] = struct{}{}
-				s.chartDateLevelCountMap.Set(date, level, mData)
-				//s.chartDateLevelCountMap[date][level][word] = struct{}{}
+				wordMapNew[word] = struct{}{}
+				s.chartDateLevelCountMap.Set(date, level, wordMapNew)
 			}
 		}
 	}
